feat: 회원 수정, 삭제 수정
회원 정보를 수정하거나 회원 삭제는 본인과 관리자만 사용 가능합니다.

회원 정보 수정은 다른 사람이 사용 중인 이메일만 사용이 불가능합니다.

+++ src/main/java/com/personal/movie/service/MemberService.java
@@ -7,18 +7,25 @@ import com.personal.movie.dto.response.MemberResponse;
 import com.personal.movie.exception.CustomException;
 import com.personal.movie.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class MemberService {
 
     private final MemberRepository memberRepository;
+    private final PasswordEncoder passwordEncoder;
     private final AuthService authService;
 
     public MemberResponse deleteMember(Long memberId) {
         Member member = memberRepository.findById(memberId)
             .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
+
+        authService.checkAuthorization(member);
+
         memberRepository.deleteById(memberId);
 
         return MemberResponse.fromEntity(member);
@@ -33,16 +40,14 @@ public class MemberService {
         Member member = memberRepository.findById(memberId)
             .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
 
-        if (memberRepository.existsByMemberName(request.getMemberName())) {
-            throw new CustomException(ErrorCode.ALREADY_EXIST_MEMBER);
-        }
-        member.updateMemberName(request.getMemberName());
+        authService.checkAuthorization(member);
 
-        if (memberRepository.existsByEmail(request.getEmail())) {
+        if (!member.getEmail().equals(request.getEmail()) && memberRepository.existsByEmail(
+            request.getEmail())) {
             throw new CustomException(ErrorCode.ALREADY_EXIST_EMAIL);
         }
         member.updateEmail(request.getEmail());
-        member.updatePassword(request.getPassword());
+        member.updatePassword(passwordEncoder.encode(request.getPassword()));
         member.updateName(request.getName());
         member.updatePhoneNumber(request.getPhoneNumber());
